restructured testing + added accuracy and precision measurements
