+++ README.md
@@ -34,8 +34,7 @@ Putting it here atm to have it in mind:
 # 📝 Tasks
 - [ ] Build "predict" -> build the class
 - [ ] Refactor code and convert to class
-- [ ] Performance analysis (computationally-wise)
+- [ ] Performance analysis (computationally-wise)
 - [ ] Performance analysis (algorithm quality-wise: e.g. precision, recall, etc -> 2 languages datasets, how making an error propagates errors)
 - [ ] EDA (Exploratory data analysis)
-- [ ] ⚠️ States -> convert to list !! ⚠️
-
+- [ ] ⚠️ States -> convert to list !! ⚠️
